fix: fix makePrivate

+++ packages/backend/src/core/NoteDeleteService.ts
@@ -77,7 +77,7 @@ export class NoteDeleteService {
 	/**
 	 * 投稿を削除します。
 	 */
-	async delete(user: MiUser, note: MiNote, deleter?: MiUser, immediate = false): Promise<void> {
+	async delete(user: MiUser, note: MiNote, deleter?: MiUser, immediate = false, opts = { makePrivate: false }): Promise<void> {
 		if (note.userId !== user.id) {
 			throw new Error(`Not deleting note ${note.id} because user ${user.id} is not the expected author ${note.userId}. This is likely a bug; please report this error to Sharkey team.`);
 		}
@@ -106,8 +106,9 @@ export class NoteDeleteService {
 		// Braces preserved to avoid merge conflicts
 		{
 			// Publish websocket deleted events
+			const publishKind = opts.makePrivate ? 'madePrivate' : 'deleted';
 			for (const note of allNotes) {
-				promises.push(this.globalEventService.publishNoteStream(note.id, 'deleted', {
+				promises.push(this.globalEventService.publishNoteStream(note.id, publishKind, {
 					id: note.id,
 					userId: note.userId,
 					body: {
@@ -180,7 +181,11 @@ export class NoteDeleteService {
 		// Don't put this in the promise array, since it needs to happen before the next section!
 		const sortedNotes = allNotes.toSorted((a, b) => b.id.localeCompare(a.id));
 		for (const note of sortedNotes) {
-			await this.notesRepository.delete({ id: note.id });
+			if (opts.makePrivate) {
+				await this.notesRepository.update({ id: note.id }, { visibility: 'specified' });
+			} else {
+				await this.notesRepository.delete({ id: note.id });
+			}
 		}
 
 		// Update the Latest Note index / following feed *after* note is deleted
@@ -192,7 +197,8 @@ export class NoteDeleteService {
 
 		// Write mod log
 		if (deleter && (user.id !== deleter.id)) {
-			promises.push(this.moderationLogService.log(deleter, 'deleteNote', {
+			const kind = opts.makePrivate ? 'makePrivateNote' : 'deleteNote';
+			promises.push(this.moderationLogService.log(deleter, kind, {
 				noteId: note.id,
 				noteUserId: note.userId,
 				noteUserUsername: user.username,
@@ -218,11 +224,11 @@ export class NoteDeleteService {
 	 * @param user 投稿者
 	 * @param note 投稿
 	 */
-	async makePrivate(user: { id: MiUser['id']; uri: MiUser['uri']; host: MiUser['host']; isBot: MiUser['isBot']; }, note: MiNote, quiet = false, deleter?: MiUser) {
-		throw new Error('TODO: makePrivate is not implemented yet');
+	async makePrivate(user: MiUser, note: MiNote, deleter?: MiUser) {
+		await this.delete(user, note, deleter, false, { makePrivate: true });
 	}
 
-	async makePrivateMany(user: { id: MiUser['id']; uri: MiUser['uri']; host: MiUser['host']; isBot: MiUser['isBot']; }, sinceDate: number, untilDate: number, countOnly = false) {
+	async makePrivateMany(user: MiUser, sinceDate: number, untilDate: number, countOnly = false) {
 		const untilId = this.idService.gen(untilDate);
 		const sinceId = this.idService.gen(sinceDate);
 		const query = this.notesRepository.createQueryBuilder()
@@ -234,8 +240,8 @@ export class NoteDeleteService {
 		if (countOnly) {
 			return query.getCount();
 		} else {
-			if (await query.getCount() > 500) {
-				throw new Error('too many notes');
+			if (await query.getCount() > 1500) {
+				throw new Error('too many notes, the max is 1500');
 			}
 			const shouldMakePrivateNotes = await query.getMany();
 			for (const note of shouldMakePrivateNotes) {
